add static cast

+++ Tests/Particles/RedistributeRTSOA/main.cpp
@@ -128,7 +128,7 @@ public:
 
             auto& particle_tile = DefineAndReturnParticleTile(lev, mfi.index(), mfi.LocalTileIndex());
             auto old_size = particle_tile.size();
-            auto new_size = old_size + host_idcpu.size();
+            auto new_size = old_size + static_cast<Long>(host_idcpu.size());
             particle_tile.resize(new_size);
 
             auto& soa = particle_tile.GetStructOfArrays();
